Split up boltdb
