+++ boltdb_batch.go
@@ -1,0 +1,50 @@
+// +build boltdb
+
+package db
+
+import "github.com/etcd-io/bbolt"
+
+// boltDBBatch stores operations internally and dumps them to BoltDB on Write().
+type boltDBBatch struct {
+	db  *BoltDB
+	ops []operation
+}
+
+// Set implements Batch.
+func (b *boltDBBatch) Set(key, value []byte) {
+	b.ops = append(b.ops, operation{opTypeSet, key, value})
+}
+
+// Delete implements Batch.
+func (b *boltDBBatch) Delete(key []byte) {
+	b.ops = append(b.ops, operation{opTypeDelete, key, nil})
+}
+
+// Write implements Batch.
+func (b *boltDBBatch) Write() error {
+	return b.db.db.Batch(func(tx *bbolt.Tx) error {
+		b := tx.Bucket(bucket)
+		for _, op := range b.ops {
+			key := nonEmptyKey(nonNilBytes(op.key))
+			switch op.opType {
+			case opTypeSet:
+				if err := b.Put(key, op.value); err != nil {
+					return err
+				}
+			case opTypeDelete:
+				if err := b.Delete(key); err != nil {
+					return err
+				}
+			}
+		}
+		return nil
+	})
+}
+
+// WriteSync implements Batch.
+func (b *boltDBBatch) WriteSync() error {
+	return b.Write()
+}
+
+// Close implements Batch.
+func (b *boltDBBatch) Close() {}
